# [CARBONDATA-3929]Improve CDC performance