@QiangCai 2020-08-24
why don't use row_v1 format in the streaming module?