@akashrn5 2020-08-25
> why don't use row_v1 format in the streaming module?
already avro is row format and same write interface it was using. I took the same with less code changes in merge command base code and as its more stable, later if we find better performance or other advantages we can try